- migration to Allegiance IGC structs: TreasureSet - detailed powerup editing - removed a warning in MDLView.cpp

+++ ICE/MDLView.cpp
@@ -61,7 +61,7 @@ BOOL CMDLView::OnInitDialog(void)
 	}
 	else
 	{
-		rawsize = fMDL.GetLength();
+		rawsize = (UINT)fMDL.GetLength();
 		rawdata = new UCHAR[rawsize];
 		fMDL.Read(rawdata,rawsize);
 		fMDL.Close();
